[TodoElement] Set the initial date and time of a newly created element to the current date and time. This makes it more convenient to create an event, but moves the event to the bottom of currently triggered events.
[TodoWidget] Added button to reset the location of the TodoWidget to the
center of the screen, as it would occasionally be lost in the outer
margins of the screen and be inaccessible.

[ClipboardWidget] Fixed clipboard load functionality (would only load
files previously).

[FractalWidget] Corrected a bug that may have throttled the rendering of
the fractal unnecessarily.

[TodoWidget] The preview window will now show the whole event instead of
the current (possibly minimized) event. The scrolling speed is slower.
Change size prompt now displays current width. TodoTools.alertBeep now
locks before the thread is started, preventing odd distortions when
multiple elements call it.